[63e244d] [UPnP] split off CUPnPRenderer to separate file
